DEVOPS-3520: add rails 8.1 support
