+++ Appraisals
@@ -1,19 +1,9 @@
-appraise 'rails7_0' do
-  gem 'activerecord', '~> 7.0'
-  gem 'sqlite3', '~> 1.4'
-end
-
-appraise 'rails7_1' do
-  gem 'activerecord', '~> 7.1'
-  gem 'sqlite3', '~> 1.4'
-end
-
-appraise 'rails7_2' do
-  gem 'activerecord', '~> 7.2'
-  gem 'sqlite3', '~> 1.4'
-end
-
 appraise 'rails8_0' do
   gem 'activerecord', '~> 8.0'
+  gem 'sqlite3', '~> 1.4'
+end
+
+appraise 'rails8_1' do
+  gem 'activerecord', '~> 8.1'
   gem 'sqlite3', '~> 1.4'
 end
