--- gemfiles/rails7_0.gemfile.lock
@@ -1,91 +1,0 @@
-PATH
-  remote: ..
-  specs:
-    calculated_attributes (1.1.0)
-      activerecord (>= 7.0.0, < 9)
-
-GEM
-  remote: https://rubygems.org/
-  specs:
-    activemodel (7.0.6)
-      activesupport (= 7.0.6)
-    activerecord (7.0.6)
-      activemodel (= 7.0.6)
-      activesupport (= 7.0.6)
-    activesupport (7.0.6)
-      concurrent-ruby (~> 1.0, >= 1.0.2)
-      i18n (>= 1.6, < 2)
-      minitest (>= 5.1)
-      tzinfo (~> 2.0)
-    appraisal (2.5.0)
-      bundler
-      rake
-      thor (>= 0.14.0)
-    ast (2.4.2)
-    concurrent-ruby (1.2.2)
-    diff-lcs (1.5.0)
-    i18n (1.14.1)
-      concurrent-ruby (~> 1.0)
-    json (2.6.3)
-    language_server-protocol (3.17.0.3)
-    mini_portile2 (2.8.8)
-    minitest (5.18.1)
-    parallel (1.23.0)
-    parser (3.2.2.3)
-      ast (~> 2.4.1)
-      racc
-    racc (1.7.1)
-    rainbow (3.1.1)
-    rake (13.0.6)
-    regexp_parser (2.8.1)
-    rexml (3.2.5)
-    rspec (3.12.0)
-      rspec-core (~> 3.12.0)
-      rspec-expectations (~> 3.12.0)
-      rspec-mocks (~> 3.12.0)
-    rspec-core (3.12.2)
-      rspec-support (~> 3.12.0)
-    rspec-expectations (3.12.3)
-      diff-lcs (>= 1.2.0, < 2.0)
-      rspec-support (~> 3.12.0)
-    rspec-mocks (3.12.6)
-      diff-lcs (>= 1.2.0, < 2.0)
-      rspec-support (~> 3.12.0)
-    rspec-support (3.12.1)
-    rubocop (1.54.2)
-      json (~> 2.3)
-      language_server-protocol (>= 3.17.0)
-      parallel (~> 1.10)
-      parser (>= 3.2.2.3)
-      rainbow (>= 2.2.2, < 4.0)
-      regexp_parser (>= 1.8, < 3.0)
-      rexml (>= 3.2.5, < 4.0)
-      rubocop-ast (>= 1.28.0, < 2.0)
-      ruby-progressbar (~> 1.7)
-      unicode-display_width (>= 2.4.0, < 3.0)
-    rubocop-ast (1.29.0)
-      parser (>= 3.2.1.0)
-    ruby-progressbar (1.13.0)
-    sqlite3 (1.6.3)
-      mini_portile2 (~> 2.8.0)
-    thor (1.2.2)
-    tzinfo (2.0.6)
-      concurrent-ruby (~> 1.0)
-    unicode-display_width (2.4.2)
-
-PLATFORMS
-  x86_64-linux
-
-DEPENDENCIES
-  activerecord (~> 7.0)
-  appraisal
-  bundler
-  calculated_attributes!
-  minitest (>= 5.18.1)
-  rake
-  rspec
-  rubocop
-  sqlite3 (~> 1.4)
-
-BUNDLED WITH
-   2.4.17
