--- gemfiles/rails7_2.gemfile
@@ -1,8 +1,0 @@
-# This file was generated by Appraisal
-
-source "https://rubygems.org"
-
-gem "activerecord", "~> 7.2"
-gem "sqlite3", "~> 1.4"
-
-gemspec path: "../"
