--- gemfiles/rails7_2.gemfile.lock
@@ -1,105 +1,0 @@
-PATH
-  remote: ..
-  specs:
-    calculated_attributes (1.1.0)
-      activerecord (>= 7.0.0, < 9)
-
-GEM
-  remote: https://rubygems.org/
-  specs:
-    activemodel (7.2.2.1)
-      activesupport (= 7.2.2.1)
-    activerecord (7.2.2.1)
-      activemodel (= 7.2.2.1)
-      activesupport (= 7.2.2.1)
-      timeout (>= 0.4.0)
-    activesupport (7.2.2.1)
-      base64
-      benchmark (>= 0.3)
-      bigdecimal
-      concurrent-ruby (~> 1.0, >= 1.3.1)
-      connection_pool (>= 2.2.5)
-      drb
-      i18n (>= 1.6, < 2)
-      logger (>= 1.4.2)
-      minitest (>= 5.1)
-      securerandom (>= 0.3)
-      tzinfo (~> 2.0, >= 2.0.5)
-    appraisal (2.5.0)
-      bundler
-      rake
-      thor (>= 0.14.0)
-    ast (2.4.2)
-    base64 (0.2.0)
-    benchmark (0.4.0)
-    bigdecimal (3.1.9)
-    concurrent-ruby (1.3.5)
-    connection_pool (2.5.0)
-    diff-lcs (1.5.0)
-    drb (2.2.1)
-    i18n (1.14.7)
-      concurrent-ruby (~> 1.0)
-    json (2.10.2)
-    language_server-protocol (3.17.0.3)
-    logger (1.6.6)
-    mini_portile2 (2.8.8)
-    minitest (5.25.5)
-    parallel (1.26.3)
-    parser (3.3.6.0)
-      ast (~> 2.4.1)
-      racc
-    racc (1.8.1)
-    rainbow (3.1.1)
-    rake (13.2.1)
-    regexp_parser (2.9.3)
-    rspec (3.12.0)
-      rspec-core (~> 3.12.0)
-      rspec-expectations (~> 3.12.0)
-      rspec-mocks (~> 3.12.0)
-    rspec-core (3.12.2)
-      rspec-support (~> 3.12.0)
-    rspec-expectations (3.12.3)
-      diff-lcs (>= 1.2.0, < 2.0)
-      rspec-support (~> 3.12.0)
-    rspec-mocks (3.12.6)
-      diff-lcs (>= 1.2.0, < 2.0)
-      rspec-support (~> 3.12.0)
-    rspec-support (3.12.1)
-    rubocop (1.69.2)
-      json (~> 2.3)
-      language_server-protocol (>= 3.17.0)
-      parallel (~> 1.10)
-      parser (>= 3.3.0.2)
-      rainbow (>= 2.2.2, < 4.0)
-      regexp_parser (>= 2.9.3, < 3.0)
-      rubocop-ast (>= 1.36.2, < 2.0)
-      ruby-progressbar (~> 1.7)
-      unicode-display_width (>= 2.4.0, < 4.0)
-    rubocop-ast (1.37.0)
-      parser (>= 3.3.1.0)
-    ruby-progressbar (1.13.0)
-    securerandom (0.4.1)
-    sqlite3 (1.6.3)
-      mini_portile2 (~> 2.8.0)
-    thor (1.3.2)
-    timeout (0.4.3)
-    tzinfo (2.0.6)
-      concurrent-ruby (~> 1.0)
-    unicode-display_width (2.6.0)
-
-PLATFORMS
-  x86_64-linux
-
-DEPENDENCIES
-  activerecord (~> 7.2)
-  appraisal
-  bundler
-  calculated_attributes!
-  minitest (>= 5.18.1)
-  rake
-  rspec
-  rubocop
-  sqlite3 (~> 1.4)
-
-BUNDLED WITH
-   2.4.17
