+++ gemfiles/rails8_0.gemfile
@@ -3,6 +3,6 @@
 source "https://rubygems.org"
 
 gem "activerecord", "~> 8.0"
-gem "sqlite3", "~> 2.6"
+gem "sqlite3", "~> 1.4"
 
 gemspec path: "../"
