+++ gemfiles/rails8_0.gemfile.lock
@@ -1,8 +1,8 @@
 PATH
   remote: ..
   specs:
-    calculated_attributes (1.1.0)
-      activerecord (>= 7.0.0, < 9)
+    calculated_attributes (1.2.0)
+      activerecord (>= 8.0.0, < 9)
 
 GEM
   remote: https://rubygems.org/
@@ -79,7 +79,7 @@ GEM
       parser (>= 3.3.1.0)
     ruby-progressbar (1.13.0)
     securerandom (0.4.1)
-    sqlite3 (2.6.0-x86_64-linux-gnu)
+    sqlite3 (1.7.3-x86_64-linux)
     thor (1.3.2)
     timeout (0.4.3)
     tzinfo (2.0.6)
@@ -99,7 +99,7 @@ DEPENDENCIES
   rake
   rspec
   rubocop
-  sqlite3 (~> 2.6)
+  sqlite3 (~> 1.4)
 
 BUNDLED WITH
    2.4.17
