+++ gemfiles/rails8_1.gemfile
@@ -2,7 +2,7 @@
 
 source "https://rubygems.org"
 
-gem "activerecord", "~> 7.0"
+gem "activerecord", "~> 8.1"
 gem "sqlite3", "~> 1.4"
 
 gemspec path: "../"
