+++ lib/calculated_attributes/version.rb
@@ -1,3 +1,3 @@
 module CalculatedAttributes
-  VERSION = '1.1.1'.freeze
+  VERSION = '1.2.0'.freeze
 end
